Restructure component's Kapitan config
With the previous structure of the component's Kapitan config we ran
into a reclass bug where nested references weren't resolved correctly if
multiple instances of the component are present with different values
for `majorVersion` (as documented for the migration to K8up 2.x).

This commit makes the major versions in the Kapitan dependency fetching
and compile instructions explicit and selects the full set of Kapitan
dependency fetching and compile instructions based on the configured
major version. This reshuffles the order in which reclass resolves
references enough to avoid running into the bug.

While this is obviously not a long-term solution for addressing this
issue, it should unblock users who're currently migrating from K8up v1.x
to v2.x using the approach outlined in the component's how-to.