Adjust K8up metrics scraping to respect exported `namespace` label
This is required in order to make the K8up metrics usable in OpenShift's
user workload monitoring (UWM). Without this, the backup namespace is
scraped as `exported_namespace` and `namespace` is the K8up namespace
for all timeseries. This makes the metrics useless for UWM users, since
UWM automatically restricts all queries with `namespace="<query source
namespace>"` which can't match backups in the source namespace, if the
backup namespace is exported as `exported_namespace`.

The commit also adjusts the component-managed alerting rules to work
with the new namespace label setup. Additionally, the commit also fixes
the alert expression for the `K8upBackupNotRunning` alert to actually
fire for individual namespaces.