Add the custom CA volume for `check` and `prune` in the `Schedule()` helper
We need to add the custom CA volume for `check`, `prune` and `backup` in
order for schedules that need a custom CA to work correctly.

This commit requires the change from the previous commit, so that
merging a custom `PruneSpec` over a `Schedule` which configures a custom
CA volume doesn't drop that config.